third_party/blink/web_tests/platform/generic/fast/filesystem/workers/sync-operations-expected.txt

[Worker] Test making multiple synchronous FileSystem operations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../resources/sync-operations.js
PASS [Worker] entries.__proto__ is Array.prototype
PASS [Worker] entries.__proto__ is Array.prototype
PASS [Worker] "/a,/b,/c,/d2,/e,/f" is "/a,/b,/c,/d2,/e,/f"
PASS [Worker] dirsCount is 3
PASS successfullyParsed is true

TEST COMPLETE
